requirements.txt
SQLAlchemy==1.3.20
strawberry-graphql==0.44.2
uvicorn==0.13.1